Fix: fixed extension appending (STABLE)
Made some mistakes using deepcopy. There was also a weird bug where the
extensions key was not being updated when passing in the config to the
compiler file. I fixed this by passing the extensions list directly to
the compile function from the user command. Also changed some filenames
for consistency